+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [0.4.0] - 2023-01-26
 - Added manual plugin installation helper.
 - Analyze the program by default, if provided, when entering the REPL with pyhidra.
 - Added language and compiler arguments to `open_program` and `run_script`.
@@ -55,7 +55,8 @@
 ## 0.1.0 - 2021-06-14
 - Initial release
 
-[Unreleased]: https://github.com/dod-cyber-crime-center/pyhidra/compare/0.3.0...HEAD
+[Unreleased]: https://github.com/dod-cyber-crime-center/pyhidra/compare/0.4.0...HEAD
+[0.4.0]: https://github.com/dod-cyber-crime-center/pyhidra/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/dod-cyber-crime-center/pyhidra/compare/0.2.1...0.3.0
 [0.2.1]: https://github.com/dod-cyber-crime-center/pyhidra/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/dod-cyber-crime-center/pyhidra/compare/0.1.5...0.2.0